Split out DataSerializableFixedId interface.
