+++ cppcache/include/geode/CacheableUndefined.hpp
@@ -22,6 +22,7 @@
 
 #include <memory>
 
+#include "internal/DataSerializableFixedId.hpp"
 #include "internal/geode_globals.hpp"
 #include "Serializable.hpp"
 #include "GeodeTypeIds.hpp"
@@ -38,7 +39,8 @@ class DataOutput;
 class Serializable;
 
 class APACHE_GEODE_EXPORT CacheableUndefined
-    : public DataSerializableFixedId_t<GeodeTypeIds::CacheableUndefined> {
+    : public internal::DataSerializableFixedId_t<
+          GeodeTypeIds::CacheableUndefined> {
  public:
   void toData(DataOutput&) const override;
