+++ cppcache/include/geode/Struct.hpp
@@ -24,6 +24,7 @@
 #include <vector>
 
 #include "internal/geode_globals.hpp"
+#include "internal/DataSerializableFixedId.hpp"
 #include "CacheableBuiltins.hpp"
 #include "StructSet.hpp"
 #include "SelectResults.hpp"
@@ -46,7 +47,7 @@ class StructSet;
  * is obtained from a Cache.
  */
 class APACHE_GEODE_EXPORT Struct
-    : public DataSerializableFixedId_t<GeodeTypeIds::Struct> {
+    : public internal::DataSerializableFixedId_t<GeodeTypeIds::Struct> {
  public:
   typedef std::vector<std::shared_ptr<Serializable>>::iterator iterator;
